requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

# Flask Framework
uvicorn>=0.18.2, <0.19.0
Flask == 2.2.2
Flask-Jsonpify >= 1.5.0
flask-requests >= 0.0.14
mysql.connector